fix: Fix the configuration of the extension when DBAL is not configured
Currently if you happen to not configure DoctrineDBAL, the extension
will append the types to the DBAL configuration resulting in an invalid
config, which will crash the booting of the Symfony app.

A typical scenario where DBAL may not be configured is if you an an
environment with in-memory tests, where DBAL is not used at all.